32-PROJECT-Texas-Hold-Em-Poker/poker/card.py
class Card:
    SUITS = ["Spades", "Clubs", "Diamonds", "Hearts"]
    RANKS = [
        "2",
        "3",
        "4",
        "5",
        "6",
        "7",
        "8",
        "9",
        "10",
        "Jack",
        "Queen",
        "King",
        "Ace",
    ]

    def __init__(self, rank, suit):
        if rank not in self.RANKS:
            raise ValueError(f"Invalid rank {rank}")
        if suit not in self.SUITS:
            raise ValueError(f"Invalid suit {suit}")

        self.rank = rank
        self.suit = suit

    def __str__(self):
        return f"{self.rank} of {self.suit}"

    def __repr__(self):
        return f"Card('{self.rank}', '{self.suit}')"
